+++ sally_theofx/src/main/java/info/kwarc/sally/theofx/ui/TheoWindow.java
@@ -43,7 +43,7 @@ import sally.Cookie;
 
 public class TheoWindow implements Runnable {
 
-	private int pid;
+	private Long pid;
 	private Logger loggr;
 	private UUID uid;
 	private int sizeY;
@@ -65,7 +65,7 @@ public class TheoWindow implements Runnable {
 
 	public static TheoWindow content;
 
-	public TheoWindow(int pid, int sizeX, int sizeY, int posX,
+	public TheoWindow(Long pid, int sizeX, int sizeY, int posX,
 			int posY, String stageTitle, String url, Cookie cookies, boolean visible) {
 		super();
 		this.pid = pid ;
@@ -92,7 +92,7 @@ public class TheoWindow implements Runnable {
 		frame.dispose();
 	}
 
-	public static TheoWindow addWindow(int pid, Integer sizeY, Integer sizeX,
+	public static TheoWindow addWindow(Long pid, Integer sizeY, Integer sizeX,
 			Integer posX, Integer posY, String stageTitle, String url, Cookie cookies, boolean visible){
 
 		try {
@@ -327,7 +327,7 @@ public class TheoWindow implements Runnable {
 	}
 
 
-	public int getPID() {
+	public Long getPID() {
 		return this.pid;
 	}
